+++ resources/views/livewire/admin/user-role/index.blade.php
@@ -20,7 +20,9 @@
                             {{-- <h5 class="card-title fw-semibold mb-4">Project Listing</h5> --}}
                         </div>
                         <div class="ms-auto">
-                            <a href="#" data-bs-toggle="modal" data-bs-target="#addRoleModal" class="btn btn-sm btn-primary text-white"><i class="fas fa-plus-square pr-4"></i>&nbsp;&nbsp; Add User Role</a>
+                            @if($admin->isAdmin)
+                                <a href="#" data-bs-toggle="modal" data-bs-target="#addRoleModal" class="btn btn-sm btn-primary text-white"><i class="fas fa-plus-square pr-4"></i>&nbsp;&nbsp; Add User Role</a>
+                            @endif
                         </div>
                     </div>
                     @if (session('message'))
@@ -38,19 +40,23 @@
                             <thead class="table-dark">
                                 <tr class="">
                                     <th class="text-secondary text-xs font-weight-semibold opacity-7 col-10">User Role</th>
-                                    <th class="text-secondary text-xs font-weight-semibold opacity-7 col-1">Action</th>
+                                    @if($admin->isAdmin)
+                                        <th class="text-secondary text-xs font-weight-semibold opacity-7 col-1">Action</th>
+                                    @endif
                                 </tr>
                             </thead>
                             <tbody>
                                 @forelse ($user_roles as $index => $user_role)
                                 <tr>
                                     <td>{{$user_role->role}}</td>
-                                    <td class="">
-                                        <div class="btn-group" role="group" aria-label="">
-                                            <a href="#" wire:click="editRole({{$user_role->id}})" data-bs-toggle="modal" data-bs-target="#editRoleModal" class="btn btn-sm btn-warning text-white"><i class="far fa-edit"></i></a>
-                                            <a href="#" wire:click="deleteRole({{$user_role->id}})" data-bs-toggle="modal" data-bs-target="#deleteRoleModal" class="btn btn-sm btn-danger text-white"><i class="fas fa-trash-alt"></i></a>
-                                        </div>
-                                    </td>
+                                    @if($admin->isAdmin)
+                                        <td class="">
+                                            <div class="btn-group" role="group" aria-label="">
+                                                <a href="#" wire:click="editRole({{$user_role->id}})" data-bs-toggle="modal" data-bs-target="#editRoleModal" class="btn btn-sm btn-warning text-white"><i class="far fa-edit"></i></a>
+                                                <a href="#" wire:click="deleteRole({{$user_role->id}})" data-bs-toggle="modal" data-bs-target="#deleteRoleModal" class="btn btn-sm btn-danger text-white"><i class="fas fa-trash-alt"></i></a>
+                                            </div>
+                                        </td>
+                                    @endif
                                 </tr>
                                 @empty
                                     <tr>
