+++ resources/views/livewire/admin/user/inactive.blade.php
@@ -39,7 +39,9 @@
                                     <th class="text-secondary text-xs font-weight-semibold opacity-7 col-4">Name</th>
                                     <th class="text-secondary text-xs font-weight-semibold opacity-7 col-4">Email Address</th>
                                     <th class="text-secondary text-xs font-weight-semibold opacity-7 col-1 text-center">Status</th>
-                                    <th class="text-secondary text-xs font-weight-semibold opacity-7 col-1">Action</th>
+                                    @if ($admin->isAdmin)
+                                        <th class="text-secondary text-xs font-weight-semibold opacity-7 col-1">Action</th>
+                                    @endif
                                 </tr>
                             </thead>
                             <tbody>
@@ -55,12 +57,14 @@
                                             <i class="fas fa-ban text-danger"></i>
                                         @endif
                                     </td>
-                                    <td class="">
-                                        <div class="btn-group" role="group" aria-label="Basic example">
-                                            <a href="#" wire:click="editUser({{$user->id}})" data-bs-toggle="modal" data-bs-target="#editUserModal" class="btn btn-sm btn-warning text-white"><i class="fas fa-pen"></i></a>
-                                            <a href="#" wire:click="deleteUser({{$user->id}})" data-bs-toggle="modal" data-bs-target="#deleteUserModal" class="btn btn-sm btn-danger text-white"><i class="fas fa-trash-alt"></i></a>
-                                        </div>
-                                    </td>
+                                    @if ($admin->isAdmin)
+                                        <td class="">
+                                            <div class="btn-group" role="group" aria-label="Basic example">
+                                                <a href="#" wire:click="editUser({{$user->id}})" data-bs-toggle="modal" data-bs-target="#editUserModal" class="btn btn-sm btn-warning text-white"><i class="fas fa-pen"></i></a>
+                                                <a href="#" wire:click="deleteUser({{$user->id}})" data-bs-toggle="modal" data-bs-target="#deleteUserModal" class="btn btn-sm btn-danger text-white"><i class="fas fa-trash-alt"></i></a>
+                                            </div>
+                                        </td>
+                                    @endif
                                 </tr>
                                 @empty
                                     <tr>
